Fix misnamed easing module
